requirements.txt
# Open requirements.txt in VS Code and replace with:
streamlit==1.31.0
PyMuPDF>=1.24.0
google-generativeai==0.3.2